requirements.txt
Django==2.1.1
psycopg2==2.7.5
pylint==2.1.1
Pillow==5.2.0
numpy==1.14.5
Keras==2.2.2
tensorflow==1.10.0
cloudinary==1.12.0
image==1.5.24